// management/frontend/src/api/schemas/workspaceMembersReadWorkspaceMembersParams.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Ego Collection Management
 * OpenAPI spec version: 0.1.0
 */

export type WorkspaceMembersReadWorkspaceMembersParams = {
/**
 * @minimum 0
 */
skip?: number;
/**
 * @maximum 500
 */
limit?: number;
job_number?: string | null;
employee_name?: string | null;
sex?: string | null;
mobile?: string | null;
employee_status?: string | null;
is_active?: boolean | null;
account_ids?: string[] | null;
business_line_ids?: string[] | null;
};