// management/frontend/src/api/zod/workspace-members/workspace-members.ts
export const workspaceMembersReadWorkspaceMembersQuerySkipDefault = 0;
export const workspaceMembersReadWorkspaceMembersQuerySkipMin = 0;

export const workspaceMembersReadWorkspaceMembersQueryLimitDefault = 20;
export const workspaceMembersReadWorkspaceMembersQueryLimitMax = 500;



export const WorkspaceMembersReadWorkspaceMembersQueryParams = zod.object({
  "skip": zod.number().min(workspaceMembersReadWorkspaceMembersQuerySkipMin).default(workspaceMembersReadWorkspaceMembersQuerySkipDefault),
  "limit": zod.number().max(workspaceMembersReadWorkspaceMembersQueryLimitMax).default(workspaceMembersReadWorkspaceMembersQueryLimitDefault),
  "job_number": zod.union([zod.string(),zod.null()]).optional(),
  "employee_name": zod.union([zod.string(),zod.null()]).optional(),
  "sex": zod.union([zod.string(),zod.null()]).optional(),
  "mobile": zod.union([zod.string(),zod.null()]).optional(),
  "employee_status": zod.union([zod.string(),zod.null()]).optional(),
  "is_active": zod.union([zod.boolean(),zod.null()]).optional(),
  "account_ids": zod.union([zod.array(zod.uuid()),zod.null()]).optional(),
  "business_line_ids": zod.union([zod.array(zod.uuid()),zod.null()]).optional()
})